+++ src/main/java/spring/springserver/domain/post/service/PostService.kt
@@ -15,6 +15,4 @@ interface PostService {
     fun updatePost(updatePostRequest: UpdatePostRequest): PostResponse
 
     fun deletePost(id: Long): DeletedPostResponse
-
-    fun purgeSoftDeletedContents()
 }
